# sdk/typescript-client/openapi.json
{
  "components": {
    "schemas": {
      "APIErrorModel": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "APIErrorModel",
        "type": "object"
      },
...
      "ChallengeCreateRequest": {
        "properties": {
          "complexity_slider": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Complexity Slider",
            "type": "number"
          },
          "iteration_window_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 86400.0,
            "title": "Iteration Window Seconds",
            "type": "integer"
          },
          "minimum_quality_threshold": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Quality Threshold",
            "type": "number"
          },
          "prompt": {
            "minLength": 10,
            "title": "Prompt",
            "type": "string"
          },
          "risk_appetite": {
            "enum": [
              "conservative",
              "balanced",
              "aggressive"
            ],
            "title": "Risk Appetite",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "prompt",
          "iteration_window_seconds",
          "minimum_quality_threshold",
          "risk_appetite",
          "complexity_slider"
        ],
        "title": "ChallengeCreateRequest",
        "type": "object"
      },
      "ChallengeResponse": {
        "properties": {
          "complexity_slider": {
            "title": "Complexity Slider",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "iteration_window_seconds": {
            "title": "Iteration Window Seconds",
            "type": "integer"
          },
          "minimum_quality_threshold": {
            "title": "Minimum Quality Threshold",
            "type": "number"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "risk_appetite": {
            "title": "Risk Appetite",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "prompt",
          "iteration_window_seconds",
          "minimum_quality_threshold",
          "risk_appetite",
          "complexity_slider",
          "created_at",
          "updated_at"
        ],
        "title": "ChallengeResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCancelResponse": {
        "properties": {
          "killed_containers": {
            "items": {
              "type": "string"
            },
            "title": "Killed Containers",
            "type": "array"
          },
          "revoked_task_ids": {
            "items": {
              "type": "string"
            },
            "title": "Revoked Task Ids",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          }
        },
        "required": [
          "run_id",
          "state",
          "killed_containers",
          "revoked_task_ids"
        ],
        "title": "RunCancelResponse",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "config_snapshot": {
            "additionalProperties": true,
            "title": "Config Snapshot",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "challenge_id",
          "state",
          "started_at",
          "ended_at",
          "config_snapshot",
          "created_at",
          "updated_at"
        ],
        "title": "RunResponse",
        "type": "object"
      },
      "RunState": {
        "enum": [
          "created",
          "running",
          "canceling",
          "completed",
          "canceled",
          "failed"
        ],
        "title": "RunState",
        "type": "string"
      },
      "RunStateTransitionRequest": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/RunState"
          }
        },
        "required": [
          "state"
        ],
        "title": "RunStateTransitionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hackathon Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/challenges": {
      "post": {
        "operationId": "create_challenge_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Challenge",
        "tags": [
          "challenges"
        ]
      }
    },
...
    "/challenges/{challenge_id}/runs/start": {
      "post": {
        "operationId": "start_run_challenges__challenge_id__runs_start_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation",
                  "details": {
                    "field": "reason"
                  },
                  "message": "Request validation failed"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIErrorModel"
                }
              }
            },
            "description": "Validation error"
          },
          "429": {
            "content": {
              "application/json": {
                "example": {
                  "code": "budget_exhausted",
                  "message": "Run budget exhausted"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIErrorModel"
                }
              }
            },
            "description": "Budget exhausted"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "code": "sandbox_failure",
                  "message": "Sandbox execution failed"
                },
                "schema": {
                  "$ref": "#/components/schemas/APIErrorModel"
                }
              }
            },
            "description": "Sandbox failure"
          }
        },
        "summary": "Start Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/challenges/{challenge_id}/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_challenges__challenge_id__runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/challenges/{challenge_id}/runs/{run_id}/state": {
      "post": {
        "operationId": "transition_run_state_challenges__challenge_id__runs__run_id__state_post",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challenge Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunStateTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition Run State",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}